feat: 관리자 주문 페이지에 각각 api 연결 및 기능 수정

+++ views/admin/order/order.js
@@ -30,7 +30,7 @@ function selectAllCheckboxes() {
 }
 
 async function deleteOrder(idList) {
-    const url = `/api/v1/orders/delete`;
+    const url = `/api/v1/orders`;
 
     try {
         const res = await fetch(url, {
@@ -38,18 +38,20 @@ async function deleteOrder(idList) {
             headers: {
                 'Content-Type': 'application/json',
             },
-            body: JSON.stringify(idList),
+            body: JSON.stringify({
+                orderIds: idList,
+            }),
         });
 
         if (res.status === 200) {
             // eslint-disable-next-line no-alert
             alert('선택하신 주문이 취소되었습니다');
         }
-    } catch(err) {
+    } catch (err) {
         // eslint-disable-next-line no-alert
         alert(`주문을 취소할 수 없습니다.`);
     }
-    
+
 }
 
 function setOrderList(date, id, addressee, orderItems, totalPrice, status) {
@@ -71,7 +73,7 @@ function setOrderList(date, id, addressee, orderItems, totalPrice, status) {
 
 async function insertOrderList() {
     // const url = './orderlistdata.json';    // 임시 데이터
-    const url = 'http://localhost:5001/api/v1/orders';
+    const url = '/api/v1/orders';
 
     try {
         const res = await fetch(url);
@@ -112,11 +114,14 @@ function orderCancel() {
         return;
     }
 
-    deleteOrder(checkedOrders);
-    while (orderListEl.firstChild) {
-        orderListEl.removeChild(orderListEl.firstChild);
+    if (window.confirm('선택한 주문을 취소하시겠습니까?')) {
+        deleteOrder(checkedOrders);
+        while (orderListEl.firstChild) {
+            orderListEl.removeChild(orderListEl.firstChild);
+        }
+        checkAll.checked = false;
+        insertOrderList();
     }
-    insertOrderList();
 }
 
 insertOrderList();
